test: move test files to `test` directory
